django-api/dropout-model/student_dropout/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
fonttools==4.53.1
htmlmin==0.1.12
idna==3.7
ImageHash==4.3.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.1
multimethod==1.12
networkx==3.3
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.1
statsmodels==0.14.2
threadpoolctl==3.5.0
tqdm==4.66.4
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visions==0.7.6
wordcloud==1.9.3
ydata-profiling==4.9.0
